Add PermitAreaItem and PermitSubjectItem
Add new supporting models for storing Permit data: PermitAreaItem and
PermitSubjectItem.  These store the data from the "areas" and "subjects"
JSON fields in a format that is easier to search.  Even though
PostgreSQL allows making special queries to JSON fields, it's still
impossible (or at least I can't find how) to search for Permits that
have a non-empty registration number in the "subjects" field, but with
these new models it's easy.

The data to PermitSubjectItem and PermitAreaItem database tables will be
created automatically when Permit is created or updated (bulk creation
is also handled) and initially for the existing Permits it's generated
with a data migration (0044_populate_permit_items_from_json).

Additionally the PermitLookupItem objects get FKs to the subject_item
and area_item that it's originating from.

+++ parkings/models/permit.py
@@ -1,5 +1,3 @@
-from itertools import chain
-
 from django.conf import settings
 from django.contrib.gis.db import models as gis_models
 from django.contrib.postgres.fields import JSONField
@@ -33,6 +31,11 @@ class PermitArea(TimestampedModelMixin):
 
     def __str__(self):
         return '{}/{}: {}'.format(self.domain.code, self.identifier, self.name)
+
+    @classmethod
+    def get_identifier_map(cls, domain, using="default"):
+        areas = cls.objects.using(using).filter(domain=domain)
+        return dict(areas.values_list("identifier", "id"))
 
 
 class PermitSeriesQuerySet(models.QuerySet):
@@ -90,10 +93,16 @@ class PermitQuerySet(models.QuerySet):
             assert isinstance(permit, Permit)
             permit.full_clean()
 
+        area_ids_cache = {}
+
         with transaction.atomic(using=self.db, savepoint=False):
             created_permits = super().bulk_create(permits, *args, **kwargs)
-            PermitLookupItem.objects.using(self.db).bulk_create(
-                chain(*(x._make_lookup_items() for x in created_permits)))
+            for permit in created_permits:
+                area_ids = area_ids_cache.get(permit.domain_id) or (
+                    area_ids_cache.setdefault(
+                        permit.domain_id,
+                        PermitArea.get_identifier_map(permit.domain, self.db)))
+                permit._create_all_items(area_ids, using=self.db, is_new=True)
             return created_permits
 
 
@@ -134,16 +143,53 @@ class Permit(TimestampedModelMixin, models.Model):
 
     def save(self, using=None, *args, **kwargs):
         self.full_clean()
+        is_new = (self.id is None)
         using = using or router.db_for_write(type(self), instance=self)
         with transaction.atomic(using=using, savepoint=False):
             super(Permit, self).save(using=using, *args, **kwargs)
-            self.lookup_items.all().using(using).delete()
-            new_lookup_items = self._make_lookup_items()
-            PermitLookupItem.objects.using(using).bulk_create(new_lookup_items)
+            self._create_all_items(using=using, is_new=is_new)
 
-    def _make_lookup_items(self):
-        for area in self.areas:
-            for subject in self.subjects:
+    def _create_all_items(self, area_ids=None, using="default", is_new=False):
+        if area_ids is None:
+            area_ids = PermitArea.get_identifier_map(self.domain, using)
+        if not is_new:
+            self.lookup_items.all().using(using).delete()
+            self.subject_items.all().using(using).delete()
+            self.area_items.all().using(using).delete()
+        subject_items = self._create_subject_items(using)
+        area_items = self._create_area_items(using, area_ids)
+        lookup_items = self._make_lookup_items(area_items, subject_items)
+        PermitLookupItem.objects.using(using).bulk_create(lookup_items)
+
+    def _create_subject_items(self, using):
+        subject_items = [
+            PermitSubjectItem(
+                permit_id=self.id,
+                registration_number=subject["registration_number"],
+                start_time=subject["start_time"],
+                end_time=subject["end_time"],
+            )
+            for subject in self.subjects
+        ]
+        PermitSubjectItem.objects.using(using).bulk_create(subject_items)
+        return subject_items
+
+    def _create_area_items(self, using, area_ids):
+        area_items = [
+            PermitAreaItem(
+                permit_id=self.id,
+                area_id=area_ids[area["area"]],
+                start_time=area["start_time"],
+                end_time=area["end_time"],
+            )
+            for area in self.areas
+        ]
+        PermitAreaItem.objects.using(using).bulk_create(area_items)
+        return area_items
+
+    def _make_lookup_items(self, area_items, subject_items):
+        for area_item in area_items:
+            for subject_item in subject_items:
                 # Calculate intersection of the area and subject ranges.
                 #
                 # An illustration of the area range [A_start,A_end],
@@ -164,19 +210,51 @@ class Permit(TimestampedModelMixin, models.Model):
                 #          I_start  I_end
                 #          |--------|
                 #
-                max_start_time = max(subject['start_time'], area['start_time'])
-                min_end_time = min(subject['end_time'], area['end_time'])
+                start_time = max(subject_item.start_time, area_item.start_time)
+                end_time = min(subject_item.end_time, area_item.end_time)
 
-                if max_start_time > min_end_time:  # Distinct ranges
+                if start_time > end_time:  # Distinct ranges
                     continue
                 yield PermitLookupItem(
                     permit=self,
+                    subject_item=subject_item,
+                    area_item=area_item,
                     registration_number=Parking.normalize_reg_num(
-                        subject['registration_number']),
-                    area=PermitArea.objects.get(identifier=area['area'], domain=self.domain),
-                    start_time=max_start_time,
-                    end_time=min_end_time
+                        subject_item.registration_number),
+                    area=area_item.area,
+                    start_time=start_time,
+                    end_time=end_time
                 )
+
+
+class PermitSubjectItem(models.Model):
+    permit = models.ForeignKey(
+        Permit, on_delete=models.CASCADE, related_name='subject_items')
+    start_time = models.DateTimeField()
+    end_time = models.DateTimeField()
+    registration_number = models.CharField(max_length=20)
+
+    def __str__(self):
+        return 'Permit {p} Subject {reg_num} ({start_time}-{end_time})'.format(
+            p=self.permit.id,
+            reg_num=self.registration_number,
+            start_time=self.start_time,
+            end_time=self.end_time)
+
+
+class PermitAreaItem(models.Model):
+    permit = models.ForeignKey(
+        Permit, on_delete=models.CASCADE, related_name='area_items')
+    start_time = models.DateTimeField()
+    end_time = models.DateTimeField()
+    area = models.ForeignKey(PermitArea, on_delete=models.PROTECT)
+
+    def __str__(self):
+        return 'Permit {p} Area {area} ({start_time}-{end_time})'.format(
+            p=self.permit.id,
+            area=self.area.identifier,
+            start_time=self.start_time,
+            end_time=self.end_time)
 
 
 class PermitLookupItemQuerySet(AnonymizableRegNumQuerySet, models.QuerySet):
@@ -206,6 +284,12 @@ class PermitLookupItemQuerySet(AnonymizableRegNumQuerySet, models.QuerySet):
 class PermitLookupItem(models.Model):
     permit = models.ForeignKey(
         Permit, related_name="lookup_items", on_delete=models.CASCADE)
+    subject_item = models.ForeignKey(
+        PermitSubjectItem, related_name="lookup_items", on_delete=models.CASCADE,
+        null=True, blank=True)
+    area_item = models.ForeignKey(
+        PermitAreaItem, related_name="lookup_items", on_delete=models.CASCADE,
+        null=True, blank=True)
     registration_number = models.CharField(max_length=20)
     area = models.ForeignKey(PermitArea, on_delete=models.PROTECT, default=None, null=True, blank=True)
     start_time = models.DateTimeField()
